API 279 r4 fix capability (#53)
* Fix Capability List of Strings with @Pattern constraints

* 3.0.0 Breaking changes to R4 Capability
- Renames to `CapabilityStatement`
- Fixes `@Pattern` constraints incorrectly applied to `List<String>`
- Changed string r4 data types with pattern restrictions for boolean and numbers, in to numbers.

+++ r4/src/main/java/gov/va/api/health/r4/api/datatypes/Age.java
@@ -5,6 +5,7 @@ import gov.va.api.health.r4.api.Fhir;
 import gov.va.api.health.r4.api.elements.Element;
 import gov.va.api.health.r4.api.elements.Extension;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.Pattern;
@@ -26,7 +27,7 @@ public final class Age implements Element {
 
   @Valid List<Extension> extension;
 
-  Double value;
+  BigDecimal value;
 
   @Pattern(regexp = "(<|<=|>=|>)")
   String comparator;
